--- groovy/assets/gcp/lang/en_us.lang
@@ -1,11 +1,0 @@
-
-# Recipemaps
-recipemap.greenhouse.name=Greenhouse
-
-# MetaTileEntities
-gcp.machine.greenhouse.name=Greenhouse
-gcp.machine.greenhouse.tooltip=Plant Growth Simulator
-
-# Tooltips
-gcp.ae2.channels.disabled=Channels are §cdisabled§7.
-gcp.ae2.channels.enabled=Channels are §cenabled§7.
