--- groovy/assets/gcp/lang/zh_cn.lang
@@ -1,7 +1,0 @@
-
-# Recipemaps
-recipemap.greenhouse.name=温室
-
-# MetaTileEntities
-gcp.machine.greenhouse.name=温室
-
