+++ groovy/assets/gregtech/lang/en_us.lang
@@ -1,3 +1,3 @@
 
 # Materials
-gcp.material.fluix=Fluix Crystal
+wf2.material.fluix=Fluix Crystal
