+++ groovy/assets/gregtech/lang/zh_cn.lang
@@ -1,3 +1,3 @@
 
 # Materials
-gcp.material.fluix=福鲁伊克斯水晶
+wf2.material.fluix=福鲁伊克斯水晶
