+++ groovy/assets/wf2/lang/en_us.lang
@@ -1,0 +1,3 @@
+# Tooltips
+wf2.ae2.channels.disabled=Channels are §cdisabled§7.
+wf2.ae2.channels.enabled=Channels are §cenabled§7.
