+++ groovy/postInit/Pvp/securitycraft.groovy
@@ -1,0 +1,505 @@
+//This is dumb as hell but I couldn't think of a better method so here we go
+def ouch = { stack -> stack.copyWithMeta(stack.getItemDamage() + 1)}
+
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_stone:*'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_sandstone:*'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_dirt:*'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_cobblestone:*'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_fence_gate:*'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_planks:*'))
+
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_stairs_oak'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_stairs_spruce'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_stairs_jungle'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_stairs_acacia'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_stairs_darkoak'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_stairs_stone'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_stairs_stone_brick'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_stairs_brick'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_stairs_nether_brick'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_stairs_quartz'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_stairs_cobblestone'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_stairs_sandstone'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_stairs_birch'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_stairs_red_sandstone'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_stairs_purpur'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_end_stone_bricks'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_red_nether_brick'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_obsidian'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_netherrack'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_end_stone'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_sea_lantern'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_bone_block'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_glass_pane'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_glowstone'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_gravel'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_stone_pressure_plate'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_wooden_pressure_plate'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_bookshelf'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_crystal_quartz_slab'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_stairs_crystal_quartz'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_sticky_piston'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_piston'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_observer'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_redstone_lamp'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_cobweb'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_grass'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_snow_block'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_ice'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_packed_ice'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_mycelium'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_clay'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_nether_wart_block'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_grass_path'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_stone_button'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_wooden_button'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_lever'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_hopper'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_cauldron'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_end_rod'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_white_glazed_terracotta'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_orange_glazed_terracotta'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_magenta_glazed_terracotta'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_light_blue_glazed_terracotta'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_yellow_glazed_terracotta'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_lime_glazed_terracotta'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_pink_glazed_terracotta'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_gray_glazed_terracotta'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_silver_glazed_terracotta'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_cyan_glazed_terracotta'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_purple_glazed_terracotta'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_blue_glazed_terracotta'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_brown_glazed_terracotta'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_green_glazed_terracotta'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_red_glazed_terracotta'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_black_glazed_terracotta'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_ladder'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_dispenser'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_dropper'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_oak_fence'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_nether_brick_fence'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_spruce_fence'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_birch_fence'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_jungle_fence'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_dark_oak_fence'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_acacia_fence'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_oak_fence_gate'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_spruce_fence_gate'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_birch_fence_gate'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_jungle_fence_gate'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_dark_oak_fence_gate'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_acacia_fence_gate'))
+
+mods.jei.ingredient.removeAndHide(item('securitycraft:universal_block_reinforcer_lvl1'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:universal_block_reinforcer_lvl2'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:universal_block_reinforcer_lvl3'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:briefcase'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:sentry'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:keypad_door_item'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:portable_tune_player'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:laser_block'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:keypad'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:keycard_reader'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_iron_trapdoor'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:cage_trap'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_iron_bars'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:keypad_chest'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:keypad_frame'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:keypad_furnace'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:sonic_security_system'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:display_case'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:keypad_trapdoor'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:scanner_trapdoor'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:door_indestructible_iron_item'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:scanner_door_item'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:lens'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:keycard_lock'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:block_pocket_wall'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:block_pocket_manager'))
+
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_walls:*'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_crystal_quartz:*'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_purpur:*'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_metals:*'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_compressed_blocks:*'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:electrified_iron_fence'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:ims'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_glass_block'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_stained_glass:*'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_wood_slabs:*'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_stone_slabs:*'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_stone_slabs2:*'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_stone_brick:*'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_mossy_cobblestone'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_brick'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_nether_brick'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_hardened_clay'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_stained_hardened_clay:*'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_logs:*'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_logs2:*'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_wool:*'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_quartz:*'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_prismarine:*'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_red_sandstone:*'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_concrete:*'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_stained_panes:*'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_carpet:*'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:reinforced_sand:*'))
+
+mods.jei.ingredient.removeAndHide(item('securitycraft:keycard_lv1'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:keycard_lv2'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:keycard_lv3'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:keycard_lv4'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:keycard_lv5'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:limited_use_keycard'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:floor_trap'))
+mods.jei.ingredient.removeAndHide(item('securitycraft:remote_access_sentry'))
+
+
+
+crafting.remove('securitycraft:panic_button')
+crafting.shapedBuilder().name('oanicbutton')
+    .output(item('securitycraft:panic_button'))
+    .matrix(' G ', 'SBS', '   ')
+    .key('B', item('minecraft:stone_button'))
+    .key('S', ore('foilSteel'))
+    .key('G', ore('paneGlass'))
+    .register()
+
+crafting.remove('securitycraft:alarm')
+crafting.shapedBuilder().name('alramclocks')
+    .output(item('securitycraft:alarm')*3)
+    .matrix(' G ', 'GBG', 'SRS')
+    .key('B', item('minecraft:glowstone_dust'))
+    .key('S', ore('foilSteel'))
+    .key('G', ore('paneGlass'))
+    .key('R', ore('boltRedAlloy'))
+    .register()
+
+crafting.remove('securitycraft:mine')
+crafting.shapedBuilder().name('la_bomba')
+    .output(item('securitycraft:mine')*4)
+    .matrix(' P ', 'IBI', ' I ')
+    .key('P', item('minecraft:stone_pressure_plate'))
+    .key('I', ore('foilSteel'))
+    .key('B', metaitem('gelled_toluene'))
+    .register()
+
+crafting.remove('securitycraft:bouncing_betty')
+crafting.shapedBuilder().name('la_bomba_boingboing')
+    .output(item('securitycraft:bouncing_betty'))
+    .matrix(' R ', 'RMR', ' S ')
+    .key('R', ore('boltLead'))
+    .key('M', item('securitycraft:mine'))
+    .key('S', ore('springSteel'))
+    .register()
+
+crafting.remove('securitycraft:claymore')
+crafting.shapedBuilder().name('claymore')
+    .output(item('securitycraft:claymore'))
+    .matrix('   ', 'DM ', 'SGS')
+    .key('D', ore('plateLead'))
+    .key('M', item('securitycraft:mine'))
+    .key('S', ore('stickSteel'))
+    .key('G', ore('screwSteel'))
+    .register()
+
+crafting.remove('securitycraft:track_mine')
+crafting.addShapeless('track_mine', item('securitycraft:track_mine'), [item('minecraft:rail'), item('securitycraft:mine'), metaitem('duct_tape').transform(ouch)])
+
+crafting.remove('securitycraft:inventory_scanner')
+crafting.shapedBuilder().name('inventory_scanner')
+    .output(item('securitycraft:inventory_scanner'))
+    .matrix('HCP', 'SRP', 'WWW')
+    .key('P', ore('plateIron'))
+    .key('H', ore('craftingToolHardHammer'))
+    .key('S', ore('paneGlass'))
+    .key('R', ore('circuitLv'))
+    .key('C', metaitem('item_filter'))
+    .key('W', ore('stoneSmooth'))
+    .register()
+
+crafting.remove('securitycraft:retinal_scanner')
+crafting.shapedBuilder().name('retinal_scanner')
+    .output(item('securitycraft:retinal_scanner'))
+    .matrix('HPP', 'SRP', 'WWW')
+    .key('P', ore('plateIron'))
+    .key('H', ore('craftingToolHardHammer'))
+    .key('S', ore('paneGlass'))
+    .key('R', ore('circuitLv'))
+    .key('W', ore('stoneSmooth'))
+    .register()
+
+crafting.remove('securitycraft:username_logger')
+crafting.shapedBuilder().name('username_logger')
+    .output(item('securitycraft:username_logger'))
+    .matrix('HPD', 'SRR', 'PGP')
+    .key('P', ore('plateIron'))
+    .key('H', ore('craftingToolHardHammer'))
+    .key('S', ore('paneGlass'))
+    .key('R', ore('circuitLv'))
+    .key('D', ore('craftingToolScrewdriver'))
+    .key('G', item('minecraft:paper'))
+    .register()
+
+
+crafting.remove('securitycraft:block_change_detector')
+crafting.shapedBuilder().name('block_change_detector')
+    .output(item('securitycraft:block_change_detector'))
+    .matrix('HPD', 'POP', 'PCP')
+    .key('P', ore('plateIron'))
+    .key('O', item('minecraft:observer'))
+    .key('H', ore('craftingToolHardHammer'))
+    .key('D', ore('craftingToolScrewdriver'))
+    .key('C', ore('circuitLv'))
+    .key('R', ore('boltRedAlloy'))
+    .register()
+
+
+crafting.remove('securitycraft:portable_radar')
+crafting.shapedBuilder().name('portable_radar')
+    .output(item('securitycraft:portable_radar'))
+    .matrix('   ', 'RSR', 'PCP')
+    .key('R', ore('stickSteel'))
+    .key('P', ore('plateSteel'))
+    .key('C', metaitem('wireGtSingleRedAlloy'))
+    .key('S', metaitem('sensor.lv'))
+    .register()
+
+crafting.remove('securitycraft:universal_block_modifier')
+crafting.shapedBuilder().name('universal_block_modifier')
+    .output(item('securitycraft:universal_block_modifier'))
+    .matrix(' GC', ' SO', 'S  ')
+    .key('S', ore('stickSteel'))
+    .key('C', ore('circuitLv'))
+    .key('O', metaitem('electric.piston.lv'))
+    .key('G', metaitem('screwTin'))
+    .register()
+
+
+crafting.remove('securitycraft:taser')
+crafting.shapedBuilder().name('taser')
+    .output(item('securitycraft:taser'))
+    .matrix('AAC', 'SSE', 'RBA')
+    .key('A', ore('foilRubber'))
+    .key('C', ore('springSteel'))
+    .key('S', metaitem('wireGtDoubleRedAlloy'))
+    .key('E', metaitem('emitter.lv'))
+    .key('R', ore('stickSteel'))
+    .key('B', metaitem('screwTinAlloy'))
+    .register()
+
+crafting.remove('securitycraft:motion_activated_light')
+crafting.shapedBuilder().name('motion_activated_light')
+    .output(item('securitycraft:motion_activated_light')*2)
+    .matrix('   ', 'GSF', 'SIS')
+    .key('G', ore('plateGlowstone'))
+    .key('S', ore('stickSteel'))
+    .key('F', ore('foilSteel'))
+    .key('I', item('minecraft:daylight_detector'))
+    .register()
+
+
+crafting.remove('securitycraft:secure_redstone_interface')
+crafting.shapedBuilder().name('secure_redstone_interface')
+    .output(item('securitycraft:secure_redstone_interface'))
+    .matrix(' T ', 'RCR', 'SSS')
+    .key('R', metaitem('wireGtSingleRedAlloy'))
+    .key('S', item('minecraft:stone_slab'))
+    .key('C', ore('circuitUlv'))
+    .key('T', item('minecraft:redstone_torch'))
+    .register()
+
+crafting.remove('securitycraft:security_camera')
+crafting.shapedBuilder().name('security_camera')
+    .output(item('securitycraft:security_camera')*4)
+    .matrix('FF ', 'RCF', ' SP')
+    .key('C', metaitem('wireGtSingleRedAlloy'))
+    .key('R', metaitem('sensor.lv'))
+    .key('F', ore('foilSteel'))
+    .key('S', ore('stickSteel'))
+    .key('P', ore('plateSteel'))
+    .register()
+
+crafting.remove('securitycraft:protecto')
+crafting.shapedBuilder().name('protecto')
+    .output(item('securitycraft:protecto'))
+    .matrix('CRC', 'ORO', 'OSO')
+    .key('C', metaitem('voltage_coil.mv'))
+    .key('R', metaitem('wireGtSingleRedAlloy'))
+    .key('O', item('minecraft:obsidian'))
+    .key('S', metaitem('sensor.mv'))
+    .register()
+
+crafting.remove('securitycraft:projector')
+crafting.shapedBuilder().name('projector')
+    .output(item('securitycraft:projector'))
+    .matrix('   ', 'LEM', 'SHS')
+    .key('L', metaitem('lensSapphire'))
+    .key('E', metaitem('emitter.mv'))
+    .key('M', metaitem('plateSilver'))
+    .key('S', ore('stickAluminium'))
+    .key('H', metaitem('hull.mv'))
+    .register()
+
+crafting.remove('securitycraft:rift_stabilizer')
+crafting.shapedBuilder().name('rift_stabilizer')
+    .output(item('securitycraft:rift_stabilizer'))
+    .matrix(' E ', 'BCB', 'PHP')
+    .key('E', item('minecraft:ender_eye'))
+    .key('B', item('gregtech:wire_coil'))
+    .key('C', metaitem('circuit.good_integrated'))
+    .key('P', metaitem('plateAluminium'))
+    .key('H', metaitem('hull.mv'))
+    .register()
+
+crafting.remove('securitycraft:trophy_system')
+crafting.shapedBuilder().name('trophy_system')
+    .output(item('securitycraft:trophy_system'))
+    .matrix('R R', 'BPB', 'SHS')
+    .key('H', metaitem('hull.mv'))
+    .key('R', ore('ringLead'))
+    .key('B', metaitem('springSmallLead'))
+    .key('P', item('securitycraft:protecto'))
+    .key('S', ore('stickAluminium'))
+    .register()
+
+crafting.remove('securitycraft:camera_monitor')
+crafting.shapedBuilder().name('camera_monitor')
+    .output(item('securitycraft:camera_monitor'))
+    .matrix('CP ', 'PMP', 'RAR')
+    .key('M', metaitem('cover.screen'))
+    .key('P', ore('plateSteel'))
+    .key('R', ore('foilRubber'))
+    .key('A', ore('paneGlass'))
+    .key('C', metaitem('sensor.lv'))
+    .register()
+
+crafting.remove('securitycraft:remote_access_mine')
+crafting.shapedBuilder().name('remote_access_mine')
+    .output(item('securitycraft:remote_access_mine'))
+    .matrix('CP ', 'PMP', 'RAR')
+    .key('M', metaitem('cover.screen'))
+    .key('P', ore('plateSteel'))
+    .key('R', ore('foilRubber'))
+    .key('A', metaitem('gemGarnetRed'))
+    .key('C', metaitem('sensor.lv'))
+    .register()
+
+crafting.remove('securitycraft:codebreaker')
+crafting.shapedBuilder().name('codebreaker')
+    .output(item('securitycraft:codebreaker'))
+    .matrix('CPG', 'PMP', 'RAR')
+    .key('M', metaitem('cover.screen'))
+    .key('P', ore('plateTitanium'))
+    .key('R', ore('foilRubber'))
+    .key('A', item('minecraft:nether_star'))
+    .key('C', metaitem('sensor.ev'))
+    .key('G', ore('circuitEv'))
+    .register()
+
+
+crafting.remove('securitycraft:smart_module')
+crafting.shapedBuilder().name('smart_module')
+    .output(item('securitycraft:smart_module'))
+    .matrix('AOB', 'PMP', 'CPD')
+    .key('P', item('minecraft:paper'))
+    .key('M', metaitem('cover.controller'))
+    .key('F', ore('foilSteel'))
+    .key('I', item('minecraft:daylight_detector'))
+    .key('A', ore('craftingToolScrewdriver'))
+    .key('B', ore('craftingToolSoftHammer'))
+    .key('C', ore('craftingToolWrench'))
+    .key('D', ore('craftingToolHardHammer'))
+    .key('O', ore('circuitLv'))
+    .register()
+
+crafting.remove('securitycraft:whitelist_module')
+crafting.shapedBuilder().name('whitelist_module')
+    .output(item('securitycraft:whitelist_module'))
+    .matrix('AOB', 'PMP', 'CPD')
+    .key('P', item('minecraft:paper'))
+    .key('M', metaitem('cover.controller'))
+    .key('F', ore('foilSteel'))
+    .key('I', item('minecraft:daylight_detector'))
+    .key('A', ore('craftingToolScrewdriver'))
+    .key('B', ore('craftingToolSoftHammer'))
+    .key('C', ore('craftingToolWrench'))
+    .key('D', ore('craftingToolHardHammer'))
+    .key('O', ore('dyeGreen'))
+    .register()
+
+crafting.remove('securitycraft:blacklist_module')
+crafting.shapedBuilder().name('blacklist_module')
+    .output(item('securitycraft:blacklist_module'))
+    .matrix('AOB', 'PMP', 'CPD')
+    .key('P', item('minecraft:paper'))
+    .key('M', metaitem('cover.controller'))
+    .key('F', ore('foilSteel'))
+    .key('I', item('minecraft:daylight_detector'))
+    .key('A', ore('craftingToolScrewdriver'))
+    .key('B', ore('craftingToolSoftHammer'))
+    .key('C', ore('craftingToolWrench'))
+    .key('D', ore('craftingToolHardHammer'))
+    .key('O', item('minecraft:coal'))
+    .register()
+
+crafting.remove('securitycraft:redstone_module')
+crafting.shapedBuilder().name('redstone_module')
+    .output(item('securitycraft:redstone_module'))
+    .matrix('AOB', 'PMP', 'CPD')
+    .key('P', item('minecraft:paper'))
+    .key('M', metaitem('cover.controller'))
+    .key('F', ore('foilSteel'))
+    .key('I', item('minecraft:daylight_detector'))
+    .key('A', ore('craftingToolScrewdriver'))
+    .key('B', ore('craftingToolSoftHammer'))
+    .key('C', ore('craftingToolWrench'))
+    .key('D', ore('craftingToolHardHammer'))
+    .key('O', item('minecraft:redstone_torch'))
+    .register()
+
+crafting.remove('securitycraft:speed_module')
+crafting.shapedBuilder().name('speed_module')
+    .output(item('securitycraft:speed_module'))
+    .matrix('AOB', 'PMP', 'CPD')
+    .key('P', item('minecraft:paper'))
+    .key('M', metaitem('cover.controller'))
+    .key('F', ore('foilSteel'))
+    .key('I', item('minecraft:daylight_detector'))
+    .key('A', ore('craftingToolScrewdriver'))
+    .key('B', ore('craftingToolSoftHammer'))
+    .key('C', ore('craftingToolWrench'))
+    .key('D', ore('craftingToolHardHammer'))
+    .key('O', item('minecraft:sugar'))
+    .register()
+
+crafting.remove('securitycraft:harming_module')
+crafting.shapedBuilder().name('harming_module')
+    .output(item('securitycraft:harming_module'))
+    .matrix('AOB', 'PMP', 'CPD')
+    .key('P', item('minecraft:paper'))
+    .key('M', metaitem('cover.controller'))
+    .key('F', ore('foilSteel'))
+    .key('I', item('minecraft:daylight_detector'))
+    .key('A', ore('craftingToolScrewdriver'))
+    .key('B', ore('craftingToolSoftHammer'))
+    .key('C', ore('craftingToolWrench'))
+    .key('D', ore('craftingToolHardHammer'))
+    .key('O', ore('toolSword'))
+    .register()
+
+crafting.remove('securitycraft:disguise_module')
+crafting.shapedBuilder().name('disguise_module')
+    .output(item('securitycraft:disguise_module'))
+    .matrix('AOB', 'PMP', 'CPD')
+    .key('P', item('minecraft:paper'))
+    .key('M', metaitem('cover.controller'))
+    .key('F', ore('foilSteel'))
+    .key('I', item('minecraft:daylight_detector'))
+    .key('A', ore('craftingToolScrewdriver'))
+    .key('B', ore('craftingToolSoftHammer'))
+    .key('C', ore('craftingToolWrench'))
+    .key('D', ore('craftingToolHardHammer'))
+    .key('O', item('minecraft:painting'))
+    .register()
