--- groovy/postInit/jei.groovy
@@ -1,3 +1,0 @@
-// Hide categories
-mods.jei.category.remove('appliedenergistics2.grinder')
-mods.jei.category.remove('appliedenergistics2.inscriber')
