--- groovy/postInit/removals.groovy
@@ -1,5 +1,0 @@
-
-crafting.remove('gregtech:electric_jetpack_advanced')
-crafting.remove('gregtech:electric_jetpack')
-crafting.remove('gregtech:fluid_jetpack')
-crafting.remove('gregtech:multi_furnace')
