--- groovy/postInit/turrets.groovy
@@ -1,67 +1,0 @@
-
-//LV
-/*
-crafting.remove('openmodularturrets:turret_base_0_vanilla')
-crafting.shapedBuilder().name('basicturret')
-    .output(item('openmodularturrets:turret_base'))
-    .matrix('ASA', 'FHF', 'PCP')
-    .key('A', metaitem('electric.piston.lv'))
-    .key('C', metaitem('circuit.electronic'))
-    .key('H', metaitem('hull.lv'))
-    .key('F', ore('plankTreatedWood'))
-    .key('P', metaitem('plateDoubleSteel'))
-    .key('S', metaitem('sensor.lv'))
-    .register()
-
-
-crafting.remove('openmodularturrets:potato_cannon_turret_vanilla')
-crafting.shapedBuilder().name('tatergun')
-    .output(item('openmodularturrets:potato_cannon_turret'))
-    .matrix(' BM', 'ORC', 'PSP')
-    .key('B', item('minecraft:potato'))
-    .key('M', item('gregtech:machine', 1625))
-    .key('O', item('openmodularturrets:intermediate_tiered', 5))
-    .key('R', item('openmodularturrets:intermediate_tiered', 10))
-    .key('C', ore('circuitUlv'))
-    .key('P', metaitem('plateSteel'))
-    .key('S', metaitem('electric.motor.lv'))
-    .register()
-
-
-crafting.remove('openmodularturrets:disposable_item_turret_vanilla')
-crafting.shapedBuilder().name('trashgun')
-    .output(item('openmodularturrets:disposable_item_turret'))
-    .matrix('GBM', 'ORC', 'PSP')
-    .key('B', ore('springSteel'))
-    .key('M', item('gregtech:machine', 1627))
-    .key('O', item('openmodularturrets:intermediate_tiered', 5))
-    .key('R', item('openmodularturrets:intermediate_tiered', 10))
-    .key('C', ore('circuitUlv'))
-    .key('P', metaitem('plateSteel'))
-    .key('S', metaitem('electric.motor.lv'))
-    .key('G', metaitem('gearIron'))
-    .register()
-
-crafting.remove('openmodularturrets:intermediate_tiered_10_vanilla')
-crafting.shapedBuilder().name('basicturret_barrel')
-    .output(item('openmodularturrets:intermediate_tiered', 10))
-    .matrix('HWS', 'BBB', 'FWF')
-    .key('H', ore('craftingToolHardHammer'))
-    .key('S', ore('craftingToolSoftHammer'))
-    .key('B', metaitem('pipeSmallFluidSteel'))
-    .key('W', ore('wireFineRedAlloy'))
-    .key('F', ore('screwSteel'))
-    .register()
-
-crafting.remove('openmodularturrets:intermediate_tiered_5_vanilla')
-crafting.shapedBuilder().name('basicturret_chamber')
-    .output(item('openmodularturrets:intermediate_tiered', 5))
-    .matrix('HFS', 'RBB', 'WWF')
-    .key('H', ore('craftingToolHardHammer'))
-    .key('S', ore('craftingToolSoftHammer'))
-    .key('B', metaitem('pipeSmallFluidSteel'))
-    .key('W', ore('wireFineRedAlloy'))
-    .key('F', ore('screwSteel'))
-    .key('R', ore('ringSteel'))
-    .register()
-    */
