+++ groovy/runConfig.json
@@ -1,6 +1,6 @@
 {
-  "packName": "GregTech Community Pack",
-  "packId": "gcp",
+  "packName": "Warfactory 2",
+  "packId": "wf2",
   "version": "1.0.0",
   "debug": false,
   "classes": [
